fix: create event mutation

+++ src/grapQL/event/event.mutation.ts
@@ -5,7 +5,7 @@ mutation createEvent(
     $date: String
     $time: String!
     $location: String!
-    $distance: Int!
+    $distance: Float!
     $when: Int!
     $meetingPoint: String!
     $description: String!
